+++ sources/execution/redirections/handle_here_doc.c
@@ -6,7 +6,7 @@
 /*   By: chrleroy <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/21 10:41:14 by chrleroy          #+#    #+#             */
-/*   Updated: 2025/06/11 19:15:01 by chrleroy         ###   ########.fr       */
+/*   Updated: 2025/06/13 12:19:51 by chrleroy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,7 +84,7 @@ bool	handle_here_doc(t_shell *minishell, t_tokn *redirections)
 	while (true)
 	{
 		line = get_user_input();
-		if (*line && !strncmp(line, limiter, len) && strlen(line) == len)
+		if (*line && !strncmp(line, limiter, len) && (int)strlen(line) == len)
 			break ;
 		expanded = expand_line(minishell, line, expansions);
 		if (*line && !expanded && *line != '$')
